refactor: 잘못된 브랜치 병합 되돌리기 및 재구축
