+++ back-end/src/main/java/com/techie/backend/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.techie.backend.user.repository;
 
-import com.techie.backend.user.User;
+import com.techie.backend.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
