+++ back-end/src/main/java/com/techie/backend/video/domain/Category.java
@@ -1,4 +1,4 @@
-package com.techie.backend.video;
+package com.techie.backend.video.domain;
 
 public enum Category {
     CLOUD,
