+++ back-end/src/main/java/com/techie/backend/watch_history/domain/WatchHistory.java
@@ -1,7 +1,7 @@
-package com.techie.backend.playlist_video;
+package com.techie.backend.watch_history.domain;
 
-import com.techie.backend.playlist.Playlist;
-import com.techie.backend.video.Video;
+import com.techie.backend.user.domain.User;
+import com.techie.backend.video.domain.Video;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -9,16 +9,16 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Getter
 @Entity
-@Table(name = "playlist_video")
-public class PlaylistVideo {
+@Table(name = "watch_histories")
+public class WatchHistory {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @JoinColumn(name = "playlist_id", nullable = false)
+    @JoinColumn(name = "user_id", nullable = false)
     @ManyToOne(fetch = FetchType.LAZY)
-    private Playlist playlist;
+    private User user;
 
     @JoinColumn(name = "video_id", nullable = false)
     @ManyToOne(fetch = FetchType.LAZY)
